게시물 접근 권한 처리 완료

+++ src/boards/boards.service.ts
@@ -4,6 +4,7 @@ import { Board } from './board.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreateBoardDto } from './dto/create-board.dto';
 import { BoardStatus } from './board-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class BoardsService {
@@ -12,30 +13,33 @@ export class BoardsService {
     private boardRepository: BoardRepository,
   ) {}
 
-  async getBoardById(id: number): Promise<Board> {
-    const found = await this.boardRepository.findOne(id);
+  async getBoardById(id: number, user: User): Promise<Board> {
+    const found = await this.boardRepository.findOne({ where: { id, userId: user.id } });
 
     if (!found) {
       throw new NotFoundException(`Can't find Board with id ${id}`);
     }
     return found;
   }
-  async getAllBoards(): Promise<Board[]> {
-    return this.boardRepository.find();
+  async getAllBoards(user: User): Promise<Board[]> {
+    const query = this.boardRepository.createQueryBuilder('board');
+    query.where('board.userId = :userId', { userId: user.id });
+    const boards = await query.getMany();
+    return boards;
   }
-  async createBoard(createBoardDto: CreateBoardDto): Promise<Board> {
-    return this.boardRepository.createBoard(createBoardDto);
+  async createBoard(createBoardDto: CreateBoardDto, user: User): Promise<Board> {
+    return this.boardRepository.createBoard(createBoardDto, user);
   }
-  async deleteBoardById(id: number): Promise<void> {
-    const result = await this.boardRepository.delete(id);
+  async deleteBoardById(id: number, user: User): Promise<void> {
+    const result = await this.boardRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Can't find Board with id ${id}`);
     }
     console.log('delete result', result);
   }
-  async updateBoardStatus(id: number, status: BoardStatus): Promise<Board> {
-    const board = await this.getBoardById(id);
+  async updateBoardStatus(id: number, status: BoardStatus, user: User): Promise<Board> {
+    const board = await this.getBoardById(id, user);
     board.status = status;
 
     await this.boardRepository.save(board);
